requirements.txt
backports.tarfile==1.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
docutils==0.21.2
idna==3.9
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
keyring==25.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
paramiko==3.4.1
parso==0.8.4
pkginfo==1.10.0
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
PyQt6==6.7.1
PyQt6-QScintilla==2.14.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyte==0.8.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ttp==0.9.5
twine==5.1.1
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
